Langchain_Prompts/chatbot.py
from langchain_openai import ChatOpenAI
from langchain_core.messages import SystemMessage, HumanMessage, AIMessage
from dotenv import load_dotenv

load_dotenv()

model = ChatOpenAI()

chat_history = [
    SystemMessage(
        content="You are a helpful assistant! You can answer questions, provide explanations, and engage in conversation."
    )
]

while True:
    user_input = input('You: ')
    chat_history.append(HumanMessage(content=user_input))
    if user_input.lower() == 'exit':
        break
    response = model.invoke(chat_history)
    chat_history.append(AIMessage(content=response.content))
    print(f'Bot: {response.content}')
    
print(chat_history)  # Print the entire chat history at the end